Move authorization-process tests to integration dir
