+++ packages/authorization-process/test/integration/createApiClient.test.ts
@@ -16,10 +16,10 @@ import {
   readLastEventByStreamId,
 } from "pagopa-interop-commons-test";
 import { authorizationApi } from "pagopa-interop-api-clients";
-import { authorizationService, postgresDB } from "./utils.js";
+import { authorizationService, postgresDB } from "../integrationUtils.js";
 
 describe("createConsumerClient", () => {
-  const organizationId: TenantId = generateId();
+  const userId: UserId = generateId();
 
   beforeAll(async () => {
     vi.useFakeTimers();
@@ -33,10 +33,11 @@ describe("createConsumerClient", () => {
   const clientSeed: authorizationApi.ClientSeed = {
     name: "Seed name",
     description: "Description",
-    members: [organizationId],
+    members: [userId],
   };
-  it("should write on event-store for the creation of a consumer client", async () => {
-    const { client } = await authorizationService.createConsumerClient(
+  it("should write on event-store for the creation of a api client", async () => {
+    const organizationId: TenantId = generateId();
+    const { client } = await authorizationService.createApiClient(
       {
         clientSeed,
       },
@@ -67,9 +68,9 @@ describe("createConsumerClient", () => {
       name: clientSeed.name,
       createdAt: new Date(),
       consumerId: organizationId,
-      kind: clientKind.consumer,
+      kind: clientKind.api,
       purposes: [],
-      users: clientSeed.members.map(unsafeBrandId<UserId>),
+      users: [unsafeBrandId<UserId>(userId)],
       description: clientSeed.description,
     };
