+++ packages/authorization-process/test/integration/createProducerKeychainKey.test.ts
@@ -37,14 +37,14 @@ import {
   userWithoutSecurityPrivileges,
   organizationNotAllowedOnProducerKeychain,
   tooManyKeysPerProducerKeychain,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneClient,
   addOneProducerKeychain,
   authorizationService,
   postgresDB,
   selfcareV2Client,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
 describe("createProducerKeychainKey", () => {
   const producerId: TenantId = generateId();
