+++ packages/authorization-process/test/integration/removeClientPurpose.test.ts
@@ -9,40 +9,43 @@ import {
 } from "pagopa-interop-commons-test";
 import {
   Client,
-  ClientUserDeletedV2,
-  UserId,
+  ClientPurposeRemovedV2,
+  PurposeId,
   generateId,
   toClientV2,
+  ClientKind,
+  TenantId,
 } from "pagopa-interop-models";
 import {
   clientNotFound,
   organizationNotAllowedOnClient,
-  clientUserIdNotFound,
-} from "../src/model/domain/errors.js";
+  clientKindNotAllowed,
+  purposeNotFound,
+} from "../../src/model/domain/errors.js";
 import {
   addOneClient,
   authorizationService,
   readLastAuthorizationEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("remove client user", () => {
-  it("should write on event-store for removing a user from a client", async () => {
+describe("remove client purpose", () => {
+  it("should write on event-store for removing a purpose from a client", async () => {
     const mockConsumer = getMockTenant();
-    const userIdToRemove: UserId = generateId();
-    const userIdToNotRemove: UserId = generateId();
+    const purposeIdToRemove: PurposeId = generateId();
+    const purposeIdToNotRemove: PurposeId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      users: [userIdToRemove, userIdToNotRemove],
+      purposes: [purposeIdToRemove, purposeIdToNotRemove],
     };
 
     await addOneClient(mockClient);
 
-    await authorizationService.removeClientUser(
+    await authorizationService.removeClientPurpose(
       {
         clientId: mockClient.id,
-        userIdToRemove,
+        purposeIdToRemove,
       },
       getMockContext({ authData: getMockAuthData(mockConsumer.id) })
     );
@@ -52,89 +55,107 @@ describe("remove client user", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockClient.id,
       version: "1",
-      type: "ClientUserDeleted",
+      type: "ClientPurposeRemoved",
       event_version: 2,
     });
 
     const writtenPayload = decodeProtobufPayload({
-      messageType: ClientUserDeletedV2,
+      messageType: ClientPurposeRemovedV2,
       payload: writtenEvent.data,
     });
 
     expect(writtenPayload).toEqual({
-      userId: userIdToRemove,
-      client: toClientV2({ ...mockClient, users: [userIdToNotRemove] }),
+      purposeId: purposeIdToRemove,
+      client: toClientV2({ ...mockClient, purposes: [purposeIdToNotRemove] }),
     });
   });
   it("should throw clientNotFound if the client doesn't exist", async () => {
     const mockConsumer = getMockTenant();
-    const userIdToRemove: UserId = generateId();
+    const purposeIdToRemove: PurposeId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      users: [userIdToRemove],
+      purposes: [purposeIdToRemove],
     };
 
     await addOneClient(getMockClient());
 
     expect(
-      authorizationService.removeClientUser(
+      authorizationService.removeClientPurpose(
         {
           clientId: mockClient.id,
-          userIdToRemove,
+          purposeIdToRemove,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer.id) })
       )
     ).rejects.toThrowError(clientNotFound(mockClient.id));
   });
-  it("should throw clientUserNotFound if the user isn't related to that client", async () => {
+  it.skip("should throw purposeNotFound if that purposeId is not related to that client", async () => {
     const mockConsumer = getMockTenant();
-    const notExistingUserId: UserId = generateId();
-    const userIdToNotRemove: UserId = generateId();
+    const notExistingPurposeId: PurposeId = generateId();
+    const purposeIdToNotRemove: PurposeId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      users: [userIdToNotRemove],
+      purposes: [purposeIdToNotRemove],
     };
 
     await addOneClient(mockClient);
 
     expect(
-      authorizationService.removeClientUser(
+      authorizationService.removeClientPurpose(
         {
           clientId: mockClient.id,
-          userIdToRemove: notExistingUserId,
+          purposeIdToRemove: notExistingPurposeId,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer.id) })
       )
-    ).rejects.toThrowError(
-      clientUserIdNotFound(notExistingUserId, mockClient.id)
-    );
+    ).rejects.toThrowError(purposeNotFound(notExistingPurposeId));
   });
   it("should throw organizationNotAllowedOnClient if the requester is not the consumer", async () => {
     const mockConsumer1 = getMockTenant();
     const mockConsumer2 = getMockTenant();
-    const userIdToRemove: UserId = generateId();
+    const purposeIdToRemove: PurposeId = generateId();
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer1.id,
-      users: [userIdToRemove],
+      purposes: [purposeIdToRemove],
     };
 
     await addOneClient(mockClient);
 
     expect(
-      authorizationService.removeClientUser(
+      authorizationService.removeClientPurpose(
         {
           clientId: mockClient.id,
-          userIdToRemove,
+          purposeIdToRemove,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer2.id) })
       )
     ).rejects.toThrowError(
       organizationNotAllowedOnClient(mockConsumer2.id, mockClient.id)
     );
+  });
+  it("should throw clientKindNotAllowed if the requester is the client api", async () => {
+    const mockClient: Client = {
+      ...getMockClient(),
+      kind: ClientKind.Enum.Api,
+    };
+
+    const organizationId: TenantId = generateId();
+
+    await addOneClient(mockClient);
+
+    expect(
+      authorizationService.removeClientPurpose(
+        {
+          clientId: mockClient.id,
+          purposeIdToRemove: generateId(),
+        },
+        getMockContext({ authData: getMockAuthData(organizationId) })
+      )
+    ).rejects.toThrowError(clientKindNotAllowed(mockClient.id));
   });
 });
