+++ packages/authorization-process/test/integration/removeClientUser.test.ts
@@ -9,43 +9,40 @@ import {
 } from "pagopa-interop-commons-test";
 import {
   Client,
-  ClientPurposeRemovedV2,
-  PurposeId,
+  ClientUserDeletedV2,
+  UserId,
   generateId,
   toClientV2,
-  ClientKind,
-  TenantId,
 } from "pagopa-interop-models";
 import {
   clientNotFound,
   organizationNotAllowedOnClient,
-  clientKindNotAllowed,
-  purposeNotFound,
-} from "../src/model/domain/errors.js";
+  clientUserIdNotFound,
+} from "../../src/model/domain/errors.js";
 import {
   addOneClient,
   authorizationService,
   readLastAuthorizationEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("remove client purpose", () => {
-  it("should write on event-store for removing a purpose from a client", async () => {
+describe("remove client user", () => {
+  it("should write on event-store for removing a user from a client", async () => {
     const mockConsumer = getMockTenant();
-    const purposeIdToRemove: PurposeId = generateId();
-    const purposeIdToNotRemove: PurposeId = generateId();
+    const userIdToRemove: UserId = generateId();
+    const userIdToNotRemove: UserId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      purposes: [purposeIdToRemove, purposeIdToNotRemove],
+      users: [userIdToRemove, userIdToNotRemove],
     };
 
     await addOneClient(mockClient);
 
-    await authorizationService.removeClientPurpose(
+    await authorizationService.removeClientUser(
       {
         clientId: mockClient.id,
-        purposeIdToRemove,
+        userIdToRemove,
       },
       getMockContext({ authData: getMockAuthData(mockConsumer.id) })
     );
@@ -55,107 +52,89 @@ describe("remove client purpose", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockClient.id,
       version: "1",
-      type: "ClientPurposeRemoved",
+      type: "ClientUserDeleted",
       event_version: 2,
     });
 
     const writtenPayload = decodeProtobufPayload({
-      messageType: ClientPurposeRemovedV2,
+      messageType: ClientUserDeletedV2,
       payload: writtenEvent.data,
     });
 
     expect(writtenPayload).toEqual({
-      purposeId: purposeIdToRemove,
-      client: toClientV2({ ...mockClient, purposes: [purposeIdToNotRemove] }),
+      userId: userIdToRemove,
+      client: toClientV2({ ...mockClient, users: [userIdToNotRemove] }),
     });
   });
   it("should throw clientNotFound if the client doesn't exist", async () => {
     const mockConsumer = getMockTenant();
-    const purposeIdToRemove: PurposeId = generateId();
+    const userIdToRemove: UserId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      purposes: [purposeIdToRemove],
+      users: [userIdToRemove],
     };
 
     await addOneClient(getMockClient());
 
     expect(
-      authorizationService.removeClientPurpose(
+      authorizationService.removeClientUser(
         {
           clientId: mockClient.id,
-          purposeIdToRemove,
+          userIdToRemove,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer.id) })
       )
     ).rejects.toThrowError(clientNotFound(mockClient.id));
   });
-  it.skip("should throw purposeNotFound if that purposeId is not related to that client", async () => {
+  it("should throw clientUserNotFound if the user isn't related to that client", async () => {
     const mockConsumer = getMockTenant();
-    const notExistingPurposeId: PurposeId = generateId();
-    const purposeIdToNotRemove: PurposeId = generateId();
+    const notExistingUserId: UserId = generateId();
+    const userIdToNotRemove: UserId = generateId();
 
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer.id,
-      purposes: [purposeIdToNotRemove],
+      users: [userIdToNotRemove],
     };
 
     await addOneClient(mockClient);
 
     expect(
-      authorizationService.removeClientPurpose(
+      authorizationService.removeClientUser(
         {
           clientId: mockClient.id,
-          purposeIdToRemove: notExistingPurposeId,
+          userIdToRemove: notExistingUserId,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer.id) })
       )
-    ).rejects.toThrowError(purposeNotFound(notExistingPurposeId));
+    ).rejects.toThrowError(
+      clientUserIdNotFound(notExistingUserId, mockClient.id)
+    );
   });
   it("should throw organizationNotAllowedOnClient if the requester is not the consumer", async () => {
     const mockConsumer1 = getMockTenant();
     const mockConsumer2 = getMockTenant();
-    const purposeIdToRemove: PurposeId = generateId();
+    const userIdToRemove: UserId = generateId();
     const mockClient: Client = {
       ...getMockClient(),
       consumerId: mockConsumer1.id,
-      purposes: [purposeIdToRemove],
+      users: [userIdToRemove],
     };
 
     await addOneClient(mockClient);
 
     expect(
-      authorizationService.removeClientPurpose(
+      authorizationService.removeClientUser(
         {
           clientId: mockClient.id,
-          purposeIdToRemove,
+          userIdToRemove,
         },
         getMockContext({ authData: getMockAuthData(mockConsumer2.id) })
       )
     ).rejects.toThrowError(
       organizationNotAllowedOnClient(mockConsumer2.id, mockClient.id)
     );
-  });
-  it("should throw clientKindNotAllowed if the requester is the client api", async () => {
-    const mockClient: Client = {
-      ...getMockClient(),
-      kind: ClientKind.Enum.Api,
-    };
-
-    const organizationId: TenantId = generateId();
-
-    await addOneClient(mockClient);
-
-    expect(
-      authorizationService.removeClientPurpose(
-        {
-          clientId: mockClient.id,
-          purposeIdToRemove: generateId(),
-        },
-        getMockContext({ authData: getMockAuthData(organizationId) })
-      )
-    ).rejects.toThrowError(clientKindNotAllowed(mockClient.id));
   });
 });
